requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.18.0
pystac-client>=0.7.0
rasterio>=1.3.0
numpy>=1.26.0
Pillow>=10.0.0
matplotlib>=3.8.0
requests>=2.31.0
zhipuai>=2.1.0
python-dotenv>=1.0.0
feedparser>=6.0.0
